requirements.txt
openai>=1.30.0
streamlit>=1.35.0
gspread>=6.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26.0
python-dotenv>=1.0.0
gtts>=2.5.1
pydub>=0.25.1
requests>=2.31.0
